+++ tests/testthat/helper.R
@@ -2,11 +2,11 @@ NEW_TESTTHAT = !packageVersion("testthat") <= "0.11.0"
 
 expect_expectation_successful = function(expr, info = NULL, label = NULL) {
   if (NEW_TESTTHAT) {
-    x = tryCatch(expr, expectation = function(e) e)
-    expect_is(x, "expectation_success", info = info, label = label)
+    res = tryCatch(expr, expectation = function(e) e)
+    expect_is(res, "expectation_success", info = info, label = label)
   } else {
-    with_reporter(ListReporter(), x <- force(expr))
-    expect_true(x$passed, info = info, label = label)
+    with_reporter(ListReporter(), res <- force(expr))
+    expect_true(res$passed, info = info, label = label)
   }
 }