+++ inst/NEWS.md
@@ -1,5 +1,6 @@
 # Version 1.7.4 (2016-04-07)
 * Compatibility with the upcoming testthat version.
+* `expect_` functions now return the checked object invisibly.
 * Changed default of argument `.var.name` for assertions and `label` for
   expectations: They now default to the return value of the exported function
   `vname` (instead of missing).